fix: adjust files
