+++ src/components/data/IconWithText/index.tsx
@@ -2,12 +2,12 @@ import { Stack } from "@inubekit/stack";
 import { Text } from "@inubekit/text";
 import { tokens } from "@design/tokens";
 
-interface IconWithTextProps {
+interface IIconWithText {
   icon: JSX.Element;
   text: string;
 }
 
-const IconWithText = (props: IconWithTextProps) => {
+const IconWithText = (props: IIconWithText) => {
   const { icon, text } = props;
   return (
     <Stack gap={tokens.spacing.s075}>
