+++ src/components/modals/ActionsModal/types.ts
@@ -1,7 +1,9 @@
 import { IIconAppearance } from "@inubekit/icon";
 
-export interface IInfoModal {
+interface IInfoModal {
   infoName: string;
   infoIcon: React.ReactNode;
   appearanceIcon: IIconAppearance;
 }
+
+export type { IInfoModal };
